GP-677 refactor, move work  in Genomics Plugin,add name restrictions and schema-version restrictions
