+++ ModelCatalogueCorePlugin/ModelCatalogueCoreGrailsPlugin.groovy
@@ -293,19 +293,7 @@ Model catalogue core plugin (metadata registry)
             link controller: 'dataArchitect', action: 'getSubModelElements', params: [format: 'xlsx', report:'NHIC'], id: true
         }
 		
-		reportsRegistry.register {
-			creates asset
-			title { "Export  XML Schema Definition " }
-			type Model
-			link controller: 'gelXml', action: 'modelsToXSD',id: true
-		}
-
-		reportsRegistry.register {
-			creates asset
-			title { "Export  XML XDCT Model" }
-			type Model
-			link controller: 'gelXml', action: 'gelXmlShredderModel',id: true
-		}
+		
         reportsRegistry.register {
             creates link
             title { "Inventory Report" }
